esio: add some HPCT hooks in prep for unit tests
Incorporated HPCT into build/link process.
Still having some GNU-specific trouble (I think) with
the Fortran-to-C interface for strings.  Specifically
the _namelen parameter coming into hpct_create_scratch_dir
is the length of the string storage in Fortran, not the
length of the string as reported by Fortran's len_trim().
Need to check this out.


git-svn-id: https://svn.ices.utexas.edu/repos/pecos/esio/trunk@6819 0bd85f0b-c151-43b9-a1e9-27231104551f